+++ lib/capybara/session.rb
@@ -620,8 +620,11 @@ module Capybara
     # Save a snapshot of the page. If `Capybara.asset_host` is set it will inject `base` tag
     #   pointing to `asset_host`.
     #
-    # If invoked without arguments it will save file to `Capybara.save_and_open_page_path`
-    #   and file will be given randomly generated filename.
+    # If invoked without arguments it will save file to `Capybara.save_path`
+    #   and file will be given randomly generated filename. If invoked with a relative path
+    #   the path will be relative to `Capybara.save_path`, which is different from
+    #   the previous behavior with `Capybara.save_and_open_page_path` where the relative path was
+    #   relative to Dir.pwd
     #
     # @param [String] path  the path to where it should be saved
     # @return [String]      the path to which the file was saved
@@ -636,8 +639,11 @@ module Capybara
     #
     # Save a snapshot of the page and open it in a browser for inspection.
     #
-    # If invoked without arguments it will save file to `Capybara.save_and_open_page_path`
-    #   and file will be given randomly generated filename.
+    # If invoked without arguments it will save file to `Capybara.save_path`
+    #   and file will be given randomly generated filename. If invoked with a relative path
+    #   the path will be relative to `Capybara.save_path`, which is different from
+    #   the previous behavior with `Capybara.save_and_open_page_path` where the relative path was
+    #   relative to Dir.pwd
     #
     # @param [String] path  the path to where it should be saved
     #
@@ -650,8 +656,11 @@ module Capybara
     #
     # Save a screenshot of page.
     #
-    # If invoked without `path` argument it will save file to `Capybara.save_and_open_page_path`
-    #   and file will be given randomly generated filename.
+    # If invoked without arguments it will save file to `Capybara.save_path`
+    #   and file will be given randomly generated filename. If invoked with a relative path
+    #   the path will be relative to `Capybara.save_path`, which is different from
+    #   the previous behavior with `Capybara.save_and_open_page_path` where the relative path was
+    #   relative to Dir.pwd
     #
     # @param [String] path    the path to where it should be saved
     # @param [Hash] options   a customizable set of options
@@ -666,8 +675,11 @@ module Capybara
     #
     # Save a screenshot of the page and open it for inspection.
     #
-    # If invoked without `path` argument it will save file to `Capybara.save_and_open_page_path`
-    #   and file will be given randomly generated filename.
+    # If invoked without arguments it will save file to `Capybara.save_path`
+    #   and file will be given randomly generated filename. If invoked with a relative path
+    #   the path will be relative to `Capybara.save_path`, which is different from
+    #   the previous behavior with `Capybara.save_and_open_page_path` where the relative path was
+    #   relative to Dir.pwd
     #
     # @param [String] path    the path to where it should be saved
     # @param [Hash] options   a customizable set of options
@@ -715,15 +727,18 @@ module Capybara
     end
 
     def prepare_path(path, extension)
-      path = default_path(extension) if path.nil?
+      if Capybara.save_path || Capybara.save_and_open_page_path.nil?
+        path = File.expand_path(path || default_fn(extension), Capybara.save_path)
+      else
+        path = File.expand_path(default_fn(extension), Capybara.save_and_open_page_path) if path.nil?
+      end
       FileUtils.mkdir_p(File.dirname(path))
       path
     end
 
-    def default_path(extension)
+    def default_fn(extension)
       timestamp = Time.new.strftime("%Y%m%d%H%M%S")
       path = "capybara-#{timestamp}#{rand(10**10)}.#{extension}"
-      File.expand_path(path, Capybara.save_and_open_page_path)
     end
 
     def scopes