+++ lib/capybara/spec/session/save_page_spec.rb
@@ -7,6 +7,7 @@ Capybara::SpecHelper.spec '#save_page' do
 
   after do
     Capybara.save_and_open_page_path = nil
+    Capybara.save_path = nil
     Dir.glob("capybara-*.html").each do |file|
       FileUtils.rm(file)
     end
@@ -26,7 +27,7 @@ Capybara::SpecHelper.spec '#save_page' do
   end
 
   it "can store files in a specified directory" do
-    Capybara.save_and_open_page_path = alternative_path
+    Capybara.save_path = alternative_path
     @session.save_page
     path = Dir.glob(alternative_path + "/capybara-*.html").first
     expect(File.read(path)).to include("Another World")
@@ -37,6 +38,20 @@ Capybara::SpecHelper.spec '#save_page' do
     expect(File.read("capybara-001122.html")).to include("Another World")
   end
 
+  it "can store files in a specified directory with a given filename" do
+    Capybara.save_path = alternative_path
+    @session.save_page("capybara-001133.html")
+    path = alternative_path + "/capybara-001133.html"
+    expect(File.read(path)).to include("Another World")
+  end
+
+  it "can store files in a specified directory with a given relative filename" do
+    Capybara.save_path = alternative_path
+    @session.save_page("tmp/capybara-001144.html")
+    path = alternative_path + "/tmp/capybara-001144.html"
+    expect(File.read(path)).to include("Another World")
+  end
+
   it "returns an absolute path in pwd" do
     result = @session.save_page
     path = File.expand_path(Dir.glob("capybara-*.html").first, Dir.pwd)
@@ -44,7 +59,7 @@ Capybara::SpecHelper.spec '#save_page' do
   end
 
   it "returns an absolute path in given directory" do
-    Capybara.save_and_open_page_path = alternative_path
+    Capybara.save_path = alternative_path
     result = @session.save_page
     path = File.expand_path(Dir.glob(alternative_path + "/capybara-*.html").first, alternative_path)
     expect(result).to eq(path)
@@ -87,6 +102,22 @@ Capybara::SpecHelper.spec '#save_page' do
 
       result = File.read(path)
       expect(result).to include("Bar")
+    end
+  end
+
+  context "with deprecated save_and_open_page_path" do
+    it "can store files in a specified directory" do
+      Capybara.save_and_open_page_path = alternative_path
+      @session.save_page
+      path = Dir.glob(alternative_path + "/capybara-*.html").first
+      expect(File.read(path)).to include("Another World")
+    end
+
+    it "returns an absolute path in given directory" do
+      Capybara.save_and_open_page_path = alternative_path
+      result = @session.save_page
+      path = File.expand_path(Dir.glob(alternative_path + "/capybara-*.html").first, alternative_path)
+      expect(result).to eq(path)
     end
   end
 end